Add cargo profile to build with --profile=small
This is primarily useful in dockerfiles when trying to create small images. 

Imho, it should be possible to ship a fairly complete web stack including both dev build tools and the webserver in a 10 MB alpine linux image without feeling too constrained by what pages can be built and served by it.